Cleans up the text writer source
* Creates `v1_0` and `v1_1` directories to mirror the structure of the
  binary encoding source.
* Makes the `Encoding` trait extend `LazyEncoder` and `LazyDecoder` so
  we can homogenize all of our generics around `Encoding`.
* Stubs out an implementation of `LazyRawTextWriter_1_1`.
