+++ src/lazy/encoder/binary/v1_0/mod.rs
@@ -1,7 +1,6 @@
 use crate::lazy::encoder::binary::v1_0::writer::LazyRawBinaryWriter_1_0;
 use crate::lazy::encoder::{LazyEncoder, SymbolCreationPolicy};
 use crate::lazy::encoding::BinaryEncoding_1_0;
-use crate::write_config::WriteConfig;
 use std::io::Write;
 
 mod container_writers;
@@ -14,8 +13,4 @@ impl LazyEncoder for BinaryEncoding_1_0 {
         SymbolCreationPolicy::RequireSymbolId;
 
     type Writer<W: Write> = LazyRawBinaryWriter_1_0<W>;
-
-    fn default_write_config() -> WriteConfig<Self> {
-        WriteConfig::<Self>::new()
-    }
 }
